route

+++ routes/api.php
@@ -36,19 +36,23 @@ Route::controller(AuthController::class)->group(function () {
 });
 
 Route::controller(c_kategori::class)->middleware('auth:sanctum')->group(function () {
-    Route::post('kategori', 'store')->middleware('role:admin');
-    Route::get('kategori/{id}', 'show');
+    Route::post('admin/kategori', 'store')->middleware('role:admin');
+    Route::get('admin/kategori/{id}', 'show')->middleware('role:admin');
+    Route::get('admin/kategori', 'get')->middleware('role:admin');
+     Route::get('kategori/{id}', 'show');
     Route::get('kategori', 'get');
-	Route::post('kategori/{id}', 'put')->middleware('role:admin');
-    Route::delete('kategori/{id}', 'delete')->middleware('role:admin');
+	Route::post('admin/kategori/{id}', 'put')->middleware('role:admin');
+    Route::delete('admin/kategori/{id}', 'delete')->middleware('role:admin');
 });
 
 Route::controller(c_fasilitas::class)->middleware('auth:sanctum')->group(function () {
-    Route::post('fasilitas', 'store')->middleware('role:admin');
+    Route::post('admin/fasilitas', 'store')->middleware('role:admin');
+    Route::get('admin/fasilitas/{id}', 'show')->middleware('role:admin');
+    Route::get('admin/fasilitas', 'get')->middleware('role:admin');
     Route::get('fasilitas/{id}', 'show');
     Route::get('fasilitas', 'get');
-	Route::post('fasilitas/{id}', 'put')->middleware('role:admin');
-    Route::delete('fasilitas/{id}', 'delete')->middleware('role:admin');
+	Route::post('admin/fasilitas/{id}', 'put')->middleware('role:admin');
+    Route::delete('admin/fasilitas/{id}', 'delete')->middleware('role:admin');
 });
 
 Route::controller(c_berita_informasi::class)->middleware('auth:sanctum')->group(function () {
